nilfs2: store parent inode number in disk inode
This uses the ->i_pad field of disk inodes to store the inode number
of their parent directory, and maintains the parent inode number for
namespace operations such as creation, rename, or unlink.

Although regular files may have two or more parent directories because
of hard link, this only maintains one parent inode and igore the hard
links.

Also, the size of parent inode numbers stored in the new field is
limited to 32-bit width even though inode number may become a 64-bit
value.

So, at this moment, this is only applicable for experimental purposes.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@lab.ntt.co.jp>