nilfs2: add helper functions to get type of blocks on ifile
This adds a helper function which classifies meta-data blocks managed
under persistent object allocator of nilfs.

The persistent object allocator uses three types of blocks, a
descriptor block, a bitmap block, and an entry block, on a single
meta-data file.  The helper function calculates this type from a given
block offset of the metadata file.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@lab.ntt.co.jp>