nilfs2: add nilfs_msg() message interface
Define an own output routine to replace bare use of printk() function.
The output routine is implemented with a macro and a helper function,
which are named nilfs_msg() and __nilfs_msg(), respectively.

__nilfs_msg() formats a message like "NILFS (<device-name>):
<message>", prefixing it with a given log level, and terminates the
statement with a newline.  The "device-name" is optional to make it
available in early stages; it will be omitted if a NULL pointer is
passed to super block instance argument.  nilfs_msg() wraps
__nilfs_msg() and is removed if CONFIG_PRINTK is not set.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@lab.ntt.co.jp>